+++ website/docs/about/what-is-di.mdx
@@ -2,10 +2,6 @@
 title: What is DI ?
 ---
 
-import {BenchmarkGraph} from "./BenchmarkGraph"
-
-<BenchmarkGraph height={300} />
-
 DI (Dependency Injection) is a general technique in OOP that all about removing unconcerned dependencies from your code.
 It brings testability, maintainability, extensibility or any kind of exchangeability to your object graph.
